Reworked on database

+++ DAL/DTO/TrainingPreRequisteDTO.cs
@@ -3,7 +3,7 @@ namespace DAL.DTO
 {
     public class TrainingPreRequisteDTO
     {
-        public int TrainingId { get; set; }
+        public short TrainingId { get; set; }
         public string PreRequisiteDescription { get; set; }
     }
 }
